Separate BLAS reference wrappers
